Réorganise les variables du covid en fichiers indépendants d'un répertoire covid
Et ajoute la référence de l'aide famille
